requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
phidata>=2.0.0
speechrecognition>=3.10.0
streamlit-webrtc>=0.47.0
av>=10.0.0
groq>=0.4.0
streamlit>=1.28.0
python-dotenv>=1.0.0
phidata>=2.0.0
speechrecognition>=3.10.0
streamlit-webrtc>=0.47.0
av>=10.0.0
groq>=0.4.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
